Added gradient function and test suite
Function to calculate the gradient and magnitude images is created and working
Seems like it needs to be the opposite way round to work like expected. That
is [1, 0, -1] rather than [-1, 0, 1]. Slicing.py is tester code for putting in
Cell and Block.
test_objects will have the test suite for handcrafted examples and the
differences between skimage HoG and our HoG. Added various import statements to cell.