Main code completed, only structure and bugs to sort
Completed histogram normalisation using a block positioning array
provided when creating the cell structure. The Hog is now linked to
all the classes however the structure of the code isn't wonderful.
Although it splits things into the order it is done well.
Got rid of bugs with histogram due to shape of array being wrong from
shape[0] to shape[1]!